requirements.txt
# AI Email Assistant Dependencies

# Google APIs
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Google Gemini API
google-generativeai==0.3.2

# Web API
flask==3.0.0

# Anthropic SDK (fallback)
anthropic==0.7.8

# MCP Server
mcp==1.0.0

# Email processing
email-validator==2.1.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
